Read VCF header from stream in VCFOutFormatter.
